requirements.txt
python-dotenv==0.17.1
python-telegram-bot==13.5
requests==2.25.1